requirements.txt
torch~=1.5.0
Pillow~=6.2.1
omegaconf~=2.0.1rc7
torchvision~=0.6.0+cpu
numpy~=1.18.4
matplotlib~=3.2.1